New tables populated
Anything that isn't catered for properly by the switch - case defaults to the old 'Exercises' table.

Currently there are 7 shoulder movements that are landing there because their specific targets are a combinarion of deltoids. I'll work something out for that later - next steps are to figure out how to build a workout from here