Made the application shut down when the form closes
... cause that was really annoying me, lol.

I've also changed a couple of the more aesthetically displeasing parts of the code to be alphabetically sorted so that finding the part that I may need in future should be made easier with some nice clean and sensible code.

Speaking of which, I think I'm going to take out that massive wall of text that generates the exercises and stick it in a JSON file to be read in. That way I can add to / change it externally without having to wait 6 weeks and a day for Visual Studio to open the project on my PC.