Solution to shoulder exercises slipping through the net
There were some shoulder movements that were defaulting to the Exercises table due to their dual specific targets (Posterior and Lateral Deltoids).

My solution to this is to just duplicate them in the JSON file, one with each specific target. I'll then deal with the duplicates at the merging stage.