Glutes!
Had an epiphany after I called it a night (post 3am) last night: I was neglecting the glutes!

I don't have any exercises for them yet but I've created a table and added code to handle it as if it was there already.

Also, the tables are viewable again... must have been some weird VS bug