Starting to create workout plans
Having some fun trying to think like a PT and plan out my own workouts - however SQL is providing me with some tough love learning techniques that are slowing down the process.

That said, I have got the foundations for a decent Strengh Push day working, and hopefully the struggles from that will simplify the rest when I get time to take the other workouts on.

I need to make my Tricep data more in-depth too, as I plan to attack each of the heads of the muscle but do not provide the database with the information required to do so. Ergo, the Push Strength workout query remains incomplete.