Pull Hypertrophy
Pretty much the same as strength but re-introducing the possibility of cable workouts.

I've said it before but these are not meant to be concrete workout plans. I am just getting these set up and then I may start incorporating them as a training program, and tweak them based on my own results analysis.

Also, the way I'm setting the queries in the DO IT button is disgusting. Needs at the very minimum a big refactor, but there's probably an even better way to do it with parsing .sql files or something. I'll look into that after push and pull hypertrophy are done, and then the next stage will probably be bodypart isolation sessions. I don't plan on using those any time soon in my training (with the exception of possibly accessories (/abs)), so those are of extremely low priority.